Updates Rector DistStragey and remove spawn
