+++ qadence/ml_tools/train_utils/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from .base_trainer import BaseTrainer
 from .config_manager import ConfigManager
 from .accelerator import Accelerator
-from .strategy import DistributionStrategy
+from .distribution import Distributor
 
 # Modules to be automatically added to the qadence.ml_tools.loss namespace
-__all__ = ["BaseTrainer", "ConfigManager", "Accelerator", "DistributionStrategy"]
+__all__ = ["BaseTrainer", "ConfigManager", "Accelerator", "Distributor"]
